+++ Components/ExtendibleHashing/GeneralHeader.h
@@ -1,0 +1,11 @@
+#ifndef GLOBAL_HEADER_H
+#define GLOBAL_HEADER_H
+
+#include <bits/stdc++.h>
+#include "Juego.h"
+#include "Bucket.h"
+#include "GeneralFunct.h"
+#include "Extendible.h"
+
+
+#endif
