+++ Components/ExtendibleHashing/Makefile
@@ -1,0 +1,12 @@
+CXX = -g++
+
+CXXFLAG = -pedantic-errors -Wall -Wextra -Werror
+
+all: main
+	./main.o
+
+main: 
+	$(CXX) -o main.o main.cpp
+
+clear:
+	rm -rf main.o data.dat ind.dat
